Remove old repo url (#167)
* Remove old repo url

* Clean up

* Clean up

+++ src/main/java/org/cts/op/transformation/grids/IGNGeographicGrid.java
@@ -268,7 +268,7 @@ public class IGNGeographicGrid extends GeographicGrid {
             }
         }
         // decimal part size --> scale
-        scale = (int) Math.rint(Math.pow(10.0, (double) nbdec));
+        scale = (int) Math.rint(Math.pow(10.0, nbdec));
         extent = new GeographicExtent("GG", y0, yL, x0, xL, modulo);
     }
 }
